Roll farm: don't delete old RC if it has replicas desired
This fixes a bug where the roll farm would delete an old RC even if it
didn't have its replicas desired count reduced to zero. The intention of
the code was to send a low urgency alert for a human operator to clean
up the old RC, however it was unconditionally deleting the old RC
shortly after sending the alert.