Wait until old RC has zero current pods before deleting
The reason for this is to accomodate a certain case of node transfers:
when a node "owned" by the old RC becomes ineligible. RCs are
responsible for cleaning up ineligible nodes if their replicas desired
count is 0, even if they are disabled. The rolling update loop needs to
wait for that to occur before deleting the old RC.

This change will also apply in the future when switch to a rollout
strategy that doesn't involve disabling an RC at all (which is enabled
by using a scheduler that allocates a different set of nodes to each
RC).