node transfers: consider scheduling successful even if conflict
At a high level a node transfer's goal is to replace an ineligible node
within a pod cluster with an eligible one. This is somewhat difficult to
accomplish when there might be 2 RCs serving the same pod cluster while
it is under a rolling update.

The challenge arises after allocating a new node: another RC might race
to schedule that node with the one doing the node transfer. Prior to
this commit the node transfer would be considered unsuccessful in this
case and it would be canceled, with another node (eventually) allocated.

However this results in a permanent increase in the size of the cluster.
As a result, this commit considers the scheduleWithSession() function
to be successful as long as the session didn't die. A conflict on the
intent key with another RC is no longer considered an error.