make rolling updates not count ineligible nodes as health
rolling updates perform calculations for how many nodes can be
"transferred" from the old RC to the new RC. This calculation is based
off of the "minimum health" guarantee the update is supposed to provide
as well as the number of nodes that are updated and healthy on the new
and old RCs.

Now that node transfers can be performed by a disabled RC, the rolling
update must anticipate that a node might be transferred (and therefore
unscheduled) from underneath it, so it needs to check the set of
eligible nodes as part of its computation if the RC's strategy is
dynamic.