Don't delete RC status until node transfer is successful
If a node transfer fails for any reason (e.g. consul goes unavailable),
we don't want to delete the RC status record because that means
forgetting the node that we allocated for the transfer, which means the
next loop attempting the transferwill allocate another node, permanently
increasing the size of the cluster

This commit preserves the RC status record until the transfer is
completed. This change has some side effects:

1) No rollback audit record is ever created for node transfers, because
there is now no notion of a rollback. We will probably add it back at
some point by using a timeout on the health of the new pod