simplify node transfers
This commit simplifies node transfers:

No locks are required when scheduling the new pod in a node transfer
anymore. The reason we used locks before is because they are deleted
automatically by consul when the session tied to them expires, which was
necessary because we weren't labeling the new pod, so it was a safety
mechanism to make sure if the RC server died, the new pod wouldn't
remain permanently. This commit avoids the issue by labeling the pod,
which means that RCs will know about them and therefore be capable of
removing them even if the server dies and the RC is handled by another
server.

This also means that we don't need to unlock the intent key at the end
of the transfer, or write labels for the pod, which sidesteps the
possibility of a race condition with another RC during a rolling update.

This change means that the "current pods" count for an RC will be
inflated by 1 when a node transfer is in progress, because the new node
and the old node are both counted. This means that we need special care
in a few places in the system.

Rolling updates already ignore "ineligible" nodes when computing health
counts, which means that only the new node should be counted and not the
old node.

In this commit we make the main RC loop skip over the old and new nodes
of a transfer in the removePods() step. This means that it will wait to
remove any pods until the transfer is over.