rc farm: acquire node transfer lock before starting node transfer
This commit makes the main RC loop acquire the RC's "node transfer lock"
before attempting to start a node transfer. If the lock is held, it will
not start a new node transfer. This allows external systems to prevent
an RC from performing a node transfer.

This is useful for example in situations where an operator wishes to
remove a specific node from a replication controller (and decrease the
size of the RC by 1). The operator can acquire the lock in order to
prevent the RC from performing a node transfer, then remove the node
from the scheduler backing the EligibleNodes() RPC, and then decrease
the replica count of the RC which will cause it to remove the node
(because it prefers to uninstall ineligible nodes).

Without the lock, there is some probability that the RC will wake up on
its own between the scheduler change and the replica count change, and a
node transfer will be started.