Roll farm: account for node transfers when enabling RCs
The roll farm has a sanity check before it enables an RC that makes sure
that the number of current pods it has is equal to its replica count.
This sanity check is meant to prevent catastrophe where (for example) an
RC could be disabled and have 0 replicas_desired and 3 current pods. In
this case, enabling it will cause those 3 pods to be unscheduled.

However, node transfers can cause an RC to have an extra "current" node
because it waits for the pod to be healthy on the new node before
removing the pod from the old node. This commit makes the RU farm check
the RC's status to see if a node transfer is happening in the case where
the current count is 1 greater than the replica count