p2-rm: allow flag to check for missing RC
Typically p2-rm assumes that a pod either does not have a
replication_contoller_id label (and thus is safe to unschedule) or it
does have one, and decrementing the replica count on the managing RC is
the proper course of action.

However in practice a pod will occasionally end up in a state where it
has the replication_controller_id label but no RC with that ID exists,
which means that we should use the manual method of removing the pod.
This commit adds a --remove-orphan flag which will make p2-rm do an
extra check that the replication controller ID returns actually exists

Additionally this commit makes p2-rm remove the labels for any pods it
deletes