Fix disable/enable tests that were failing on travis
These tests execute bash scripts included in the repo and confirm that
the expected output came back. However travis started injecting a
warning into the output. Rather than debug why this is happening, we can
just check that the output we expect is a substring of the output we
receive