node transfers: check for unused allocations before allocating
Node transfers used to call AllocateNodes() on the scheduler to find a
node to transfer the ineligible pod to unconditionally. This commit
makes the RC first check if there is an eligible node that is already
unused (not scheduled with the pod the RC manages) and uses it instead
if it's empty.

This can prevent situations where AllocateNodes() returns a timeout
error for example, but the scheduler ends up allocating a node.
Additionally it can handle cases where a node is allocated out of band
in advance of the node transfer