Fix possible race in daemon set tests
There's a goroutine that fails the test if a session renewal error is
encountered. However the session is destroyed in a defer, which means
that there's a race where possibly a renewal error will occur after the
destroy.

This commit ensures that the goroutine reading off the error renewal
channel is shut down before the session is destroyed