Add retry with backoff to finishTransfer() call
Before this change, if there was an error building the final node
transfer transaction, whether due to a network partition or otherwise,
the node transfer would fail but would not rollback. This prevented
future node transfers from occurring.

This change adds a retry around the finishTransfer() call so that
something like a temporary network partition will not stop a node
transfer. In the event that the node transfer still fails after the
retries, there is likely an underlying problem so we now alert.