roll farm: Treat failure to delete the old RC as RU failure
An intermittent HTTP error can currently cause an RU to "finish" and
delete the RU record without cleaning up the old RC. This means that
another RU cannot be started until the old one is restarted and allowed
to complete successfully.

This commit just propagates errors from cleanupOldRC all the way up the
chain so that the RU doesn't get deleted, so the whole RU will be
retried automatically.