Extend manifest to allow for a pod-wide artifact registry override
This is a clumsy change because the artifact registry is set at Preparer startup
time. In this change, I chose cache the HTTP client and use it later on if the
manifest we are installing desires an override. All errors are recoverable by
falling back to the default artifact registry.