Ensure runit supervise directory has 755 permissions
The runit supervise directory contains information that is useful to
pods that wish to customize their deployment strategy, for example if
they wish to send the pod process a SIGHUP from a bin/post-activate
script.

If the supervise directory does not exist ahead of time, runsvdir will
create it with permissions of 700, which is too restrictive to allow a
pod to access the contents of the directory.

This commit initializes the directory ahead of time with permissions of
755 which should allow reading files that are under the directory as
long as the files themselves are readable by the pod user.