Add no_halt_on_update launchable option to launchable stanza spec
When set, this option instructs P2 to skip the "sv stop" portion of the
normal manifest update process. This is useful for processes that are
designed to gracefully reload without service interruption. However, P2
does not arrange for a signal to be sent to the process, pods that need
this functionality should implement their own signal handling in
bin/post-activate.

There are certain circumstances in which the no_halt_on_update setting
will be ignored, for instance if an operator is shutting down all pods
on the host or if a pod is being uninstalled (manifest deleted from the
intent tree).