Use `sv start` not `sv restart` for no_halt_on_update launchables
Restart sends a SIGTERM before starting the service, so launchables that
have asked not to be halted when updated should use `sv start` instead.

NOTE: i'm not sure why `sv restart` is used in the common case, it
likely has something to do with the preparer being able to deploy
itself.