Detect runc settings based on OS to support centos6
Centos6 has a few limitations that require adding runtime flags to runc
invocations that start containers.

First, runc needs a tmpfs volume to store container state on. By default
this is /run, however Centos6 doesn't ship with a tmpfs mount at this
location. To deal with this, the preparer will check if /run is a
directory and if not, it will configure runc to store container state in
/dev/shm/runc.

Secondly, centos6 doesn't have the kernel functionality to support
session keyrings, so if the preparer detects it's running on centos6 it
will add the --no-new-keyring flag to all runc run calls.