Allow overriding the OS version file
It's long been expected that p2-preparer should only be run on machines
that have the /etc/redhat-release, however only recently when the
opencontainer launchable type was added to the integration test did it
become a hard requirement for the integration test to pass.

This is because the opencontainer launchable makes some decisions based
on the contents of the file. Unfortunately we don't have access to write
the /etc/redhat-release file on a travis CI node, so this commit allows
overriding the location of the file so that it's somewhere we can write
it