opencontainer: implement PostInstall() and proper cgrouping
The runc spec has a lot of configuration that a launchable might want
to customize on its own, for example setting the uid/gid that the
container should run as.

This doesn't introduce any vulnerabilities because we check the runc
configuration for settings that we care about before starting it, for
example uid/gid and capabilities.