Relocate the launchable PostInstall() invocation after env setup
This will allow a launchable's PostInstall() script to depend on
environment being written out to disk.

This change is good conceptually, but given that a single function
writes the pod-wide env dir and all of the launchable env dirs, it came
out a little ugly code-wise.

The code in this commit optimized minimal disturbance of hoist
launchable code (which doesn't make use of PostInstall()) at the expense
of cleanliness.