opencontainer: change out post-install for pre-launch
The goal of the post-install hook was to give a container a chance to
set up its config.json file just before being launched. However,
post-install is problematic because it might depend on system setup such
as environment variables which aren't available until later in the life
cycle.

This commit instead allows for a pre-launch script which runs just
before launching (after some validation of the config.json has been
performed).