Don't attempt to validate missing config.json file in Executables()
When generating runit services for an opencontainer, we were previously
requiring that a config.json file exist for validation. However,
launchables might want to create the config.json file in the
bin/pre-launch script, so only do the validation if the file exists.

We'll do the validation again just before starting the runit services.