Add config.json rewriting support to p2-exec for open containers
This is a step toward addressing some usability concerns for the
opencontainer launchable type. The opencontainer spec requires a large
config.json file to fully describe the context of the container,
including uid/gid, and bind mounts. P2 itself knows the uid/gid to set,
so it shouldn't be required for the builder of the runc artifact to do
this correctly. It also might have a sense of reasonable defaults for
bind mounts.

This commit adds functionality to p2-exec to accept commandline flags
that are specific to containers and to make modifications to a
config.json template and write out a final config.json to be launched.
It also performs some validation of the final config.json template to
enforce constraints.