Wire up opencontainer launchable to new p2-exec flags
This commit should cause the opencontainer launchable handler to
configure p2-exec properly to generate config.json from the specified
JSON template with the correct uid/gid. Additionally, it will configure
bind mounts based on the
github.com/square/p2/pkg/pods.DefaultContainerBindMountEnvVars variable
which should be set at build time.

If environment variables are present in
DefaultContainersBindMountEnvVars, the environment variables will be
looked up and validated. Validation confirms that the values of the
variables are paths to existing directories, and that the directories
are within the pod's home directory.

Environment variables are used so that an environment variable can be
set up by a P2 hook which is outside of the P2 codebase. It's up to the
hook to create the directory that will be bind mounted.

The variable containing the default environment variables to use is
expected to be set at build time because P2 exposes a suite of tools
which may or may not need to be configured to set the correct values at
runtime. It's much easier to set it once during the build process and
have it baked into all of the deployed tools.