Add integration test logic to confirm p2-exec behavior
This commit configures the integration test properly to set up a
HELLOTMPDIR environment-variable-based bind mount into the
hello-opencontainer opencontainer launchable's container. We also modify
the source code of hello to verify that it is able to see a file that is
put into the bind mounted directory outside of the container to confirm
that it can see the file and that the bind mount is working properly.

Additionally, hello verifies that it is running as the nonprivileged
user that it should be, rather than as root which is what is specified
in its config.json.template file.