Add naive retries with backoff to getMatches
Lately we've seen network errors cause this function to fail with some
frequency. This changes adds naive retries with a simple linear backoff
to reduce the return of errors due to fleeting network failures. At the
same time, it tries to not dramatically increase latency (where latency
is defined as the time it takes getMatches to return, successfully or
not).