Remove out-of-place code snippet in the pod cluster store.
There was logic in the MutatePC function where it would write labels for
the pod after doing a mutation, and if the label write fails it would
delete the pod cluster. It doesn't make sense to be touching labels at
all here, and it certainly doesn't make sense to delete the pod cluster
unless that's what the caller wanted. Looks like this was a bad
copy/paste from the Create() function, where it does make sense to
delete a pod cluster for which a label write failed.