Always use cpu.cfs_period_us of 1,000,000
It is possible to have a pod with cpu restrictions and an
unrestricted launchable. When that happens, the pod has a cgroup
cpu.cfs_period_us of 1,000,000. However, for unrestricted cpu limits, we
switch to the default period of 100,000. Having a period of 100,000
nested within a cgroup that has a period of 1,000,000 is invalid.

This change switches to always using a period of 1,000,000 for
consistency and to prevent failed writes. Those failed writes block pod
installation.