Retry node transfers on health check timeout
When the watchHealth go routine times out waiting for a new node to come
up healthy, the routine ends. In this case, we want the node transfer to
be retried with the same old and new nodes, and that is possible because
they have been persisted in the Consul replication controller status
tree. However, because the local rc.nodeTransfer struct would have a
non-nil quit channel, the replication controller did not retry because
it thought the watch health go routine was still operating in the
background. By setting the rc.nodeTransfer struct to the empty struct,
the node transfer will be correctly retried.