Rewrite node transfers
Node transfers previously did not account for the case when the newly
scheduled pod did not ever become healthy. When that happened, rolling
updates could not complete. Additionally there was a lot of complexity
in general surrounding the "in progress" state of a node transfer. State
had to be persisted in Consul, a goroutine needed to be spawned to watch
the health of the new pod, and the RC "lied" about its state to things
like rolling updates and the allocator.

Node transfers had these issues because we wanted to ensure that a
new pod was healthy before unscheduling the old. The new transfers ditch
that requirement. Instead we check the health of all of the RC's pods
and perform a transactional swap if one of two health conditions is met:
1) the ineligible node is unhealthy (so a swap will not reduce the
overall cluster health), 2) all pods are healthy (so the cluster can
tolerate one being down).

During a rolling update, the RC wishing to perform a node transfer may
not own all of the cluster's pods, so the above health check would be
insufficient. Additionally, during a rolling update there may be races
between the respective health checks of the rolling update and transfer
and the scheduling that they each perform; e.g. both check health at a
time when all pods are healthy, then both act and make two pods
unhealthy, reducing the cluster health beyond the expected minimum. To
combat this, the RC will attempt to acquire the mutation lock on its ID.
That is the same lock that rolling updates acquire to modify the RC.
This will ensure that a transfer and a rolling update do not occur
simultaneously.

If the preconditions for a transfer are not met (health and mutation
lock) the transfer will not occur. Should the RC still be scheduled to
an ineligible node on the next call of meetDesires(), the conditions
will be checked again with the goal of eventually performing the
transfer.