Allow some node transfers during rolling updates
When the old RC in an RU is only scheduled on ineligible nodes and has
fewer replicas desired than current replicas, the rolling update can not
take any action. None of the nodes are eligible for the new RC, and the
old RC is disabled. In this situation, the old RC will now be permitted
to perform a node transfer so that the cluster will gain an eligible
node with which the new RC can schedule. The RU will now be able to
ultimately finish on a set of eligible nodes.